Delete children in reversed order and add get_option_columns
Delete children in reversed order, e.g. from bottom to top in the
hierarchy. So it is easier to prevent a parent deleting the child
accidently.

Get_option_columns is important for comboboxes. If you show the
options in a combobox browser, the leafes are taskfileinfos, thus
the name of the elemen (the first column) will be shown. But you might
want to change that to another column, like column 3 (the version).
This will return the column for each level.